Moved components into components dir, added fetch call to get user access level from server.
