--- src/AuthNav.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import AuthButton from './AuthButton';
-
-const AuthNav = () => (
-  <div className="navbar-nav ml-auto">
-    <AuthButton />
-  </div>
-);
-
-export default AuthNav;
