--- src/Login.js
@@ -1,9 +1,0 @@
-import { useAuth0 } from "@auth0/auth0-react";
-
-const Login = () => {
-  const { loginWithRedirect } = useAuth0();
-
-  return loginWithRedirect();
-};
-
-export default Login;
